+++ src/config/default.docunotion.config.ts
@@ -13,8 +13,8 @@ import { standardEscapeHtmlBlockModifier } from "../plugins/EscapeHtmlBlockModif
 import { standardHeadingTransformer } from "../plugins/HeadingTransformer";
 import { standardNumberedListTransformer } from "../plugins/NumberedListTransformer";
 import { standardTableTransformer } from "../plugins/TableTransformer";
-import { IDocuNotionConfig } from "../plugins/pluginTypes";
 import { standardExternalLinkConversion } from "../plugins/externalLinks";
+import { IDocuNotionConfig } from "./configuration";
 
 const defaultConfig: IDocuNotionConfig = {
   plugins: [
