+++ src/index.ts
@@ -7,4 +7,5 @@ run();
 // for plugins to import
 
 export * as Log from "./log";
+export * from "./types";
 export * from "./plugins/pluginTypes";
